chore: create a new module moving security related services creation
