+++ core/common/connector-core/src/main/resources/META-INF/services/org.eclipse.edc.spi.system.ServiceExtension
@@ -14,7 +14,5 @@
 
 org.eclipse.edc.connector.core.CoreServicesExtension
 org.eclipse.edc.connector.core.CoreDefaultServicesExtension
-org.eclipse.edc.connector.core.SecurityDefaultServicesExtension
-org.eclipse.edc.connector.core.LocalPublicKeyDefaultExtension
 org.eclipse.edc.connector.core.discovery.DiscoveryDefaultServicesExtension
 org.eclipse.edc.connector.core.discovery.DiscoveryServicesExtension
