+++ core/common/security-core/src/test/resources/rsa_2048.pem
